# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SHIVAM ANIMES Premium Streaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (user or admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # ─── ANIME ──────────────────────────────────────────────────────────────────
  /anime:
    get:
      operationId: listAnime
      tags: [anime]
      summary: List all anime
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of anime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeListResponse"

    post:
      operationId: createAnime
      tags: [anime]
      summary: Create new anime (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnimeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anime"
...
  /anime/{id}:
    get:
      operationId: getAnime
      tags: [anime]
      summary: Get anime by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Anime details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeWithEpisodes"
        "404":
          description: Not found

    patch:
      operationId: updateAnime
      tags: [anime]
      summary: Update anime (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnimeUpdate"
      responses:
        "200":
          description: Updated anime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anime"

    delete:
      operationId: deleteAnime
      tags: [anime]
      summary: Delete anime (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── EPISODES ───────────────────────────────────────────────────────────────
  /anime/{animeId}/episodes:
    get:
      operationId: listEpisodes
      tags: [episodes]
      summary: List episodes for an anime
      parameters:
        - name: animeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Episodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Episode"

    post:
      operationId: createEpisode
      tags: [episodes]
      summary: Add episode (admin only)
      parameters:
        - name: animeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EpisodeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Episode"
  /anime/{animeId}/episodes/{episodeId}:
    patch:
      operationId: updateEpisode
      tags: [episodes]
      summary: Update episode (admin only)
      parameters:
        - name: animeId
          in: path
          required: true
          schema:
            type: integer
        - name: episodeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EpisodeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Episode"

    delete:
      operationId: deleteEpisode
      tags: [episodes]
      summary: Delete episode (admin only)
      parameters:
        - name: animeId
          in: path
          required: true
          schema:
            type: integer
        - name: episodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── STREAM TOKEN ───────────────────────────────────────────────────────────
  /stream/token/{episodeId}:
    get:
      operationId: getStreamToken
      tags: [stream]
      summary: Get encrypted temporary stream token for an episode (premium only)
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Temporary stream token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamTokenResponse"
        "403":
          description: Not premium

  # ─── USERS ──────────────────────────────────────────────────────────────────
...
  /solve-links:
    get:
      operationId: listSolveLinks
      tags: [solveLinks]
      summary: List all solve link posts
      responses:
        "200":
          description: Solve link posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SolveLink"

    post:
      operationId: createSolveLink
      tags: [solveLinks]
      summary: Create solve link post (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolveLinkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolveLink"
  /solve-links/{id}:
    delete:
      operationId: deleteSolveLink
      tags: [solveLinks]
      summary: Delete solve link post (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ANALYTICS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        deviceId:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, role, isPremium]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [user, admin]
        isPremium:
          type: boolean
        premiumUntil:
          type: ["string", "null"]
        isSuspended:
          type: boolean
        deviceId:
          type: ["string", "null"]
        lastLogin:
          type: ["string", "null"]
        lastLoginIp:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Anime:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        releaseYear:
          type: ["integer", "null"]
        bannerImage:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        trailerUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [ongoing, completed]
        episodeCount:
          type: integer
        viewCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    AnimeInput:
      type: object
      required: [title, status]
      properties:
        title:
          type: string
        description:
          type: string
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        releaseYear:
          type: ["integer", "null"]
        bannerImage:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        trailerUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [ongoing, completed]
    AnimeUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        releaseYear:
          type: ["integer", "null"]
        bannerImage:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        trailerUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [ongoing, completed]
    AnimeListResponse:
      type: object
      required: [anime, total, page, limit]
      properties:
        anime:
          type: array
          items:
            $ref: "#/components/schemas/Anime"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AnimeWithEpisodes:
      type: object
      required: [id, title, status, createdAt, episodes]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        releaseYear:
          type: ["integer", "null"]
        bannerImage:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        trailerUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [ongoing, completed]
        episodeCount:
          type: integer
        viewCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/Episode"
    Episode:
      type: object
      required: [id, animeId, number, title, createdAt]
      properties:
        id:
          type: integer
        animeId:
          type: integer
        number:
          type: integer
        title:
          type: string
        thumbnail:
          type: ["string", "null"]
        createdAt:
          type: string
    EpisodeInput:
      type: object
      required: [number, title]
      properties:
        number:
          type: integer
        title:
          type: string
        streamUrl:
          type: string
        thumbnail:
          type: ["string", "null"]
    EpisodeUpdate:
      type: object
      properties:
        number:
          type: ["integer", "null"]
        title:
          type: ["string", "null"]
        streamUrl:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
    StreamTokenResponse:
      type: object
      required: [token, expiresAt]
      properties:
        token:
          type: string
        expiresAt:
          type: string
    SolveLink:
      type: object
      required: [id, title, shortenerUrl, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        shortenerUrl:
          type: string
        isTrending:
          type: boolean
        createdAt:
          type: string
    SolveLinkInput:
      type: object
      required: [title, shortenerUrl]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        shortenerUrl:
          type: string
        isTrending:
          type: boolean
...