# lib/api-zod/src/generated/types/animeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SHIVAM ANIMES Premium Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AnimeInputStatus } from './animeInputStatus';

export interface AnimeInput {
  title: string;
  description?: string;
  genres?: string[];
  tags?: string[];
  /** @nullable */
  releaseYear?: number | null;
  /** @nullable */
  bannerImage?: string | null;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  trailerUrl?: string | null;
  status: AnimeInputStatus;
}