# lib/api-zod/src/generated/types/animeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SHIVAM ANIMES Premium Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AnimeUpdateStatus } from './animeUpdateStatus';

export interface AnimeUpdate {
  title?: string;
  /** @nullable */
  description?: string | null;
  genres?: string[];
  tags?: string[];
  /** @nullable */
  releaseYear?: number | null;
  /** @nullable */
  bannerImage?: string | null;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  trailerUrl?: string | null;
  status?: AnimeUpdateStatus;
}