# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SHIVAM ANIMES Premium Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  username: string;
  role: UserRole;
  isPremium: boolean;
  /** @nullable */
  premiumUntil?: string | null;
  isSuspended?: boolean;
  /** @nullable */
  deviceId?: string | null;
  /** @nullable */
  lastLogin?: string | null;
  /** @nullable */
  lastLoginIp?: string | null;
  createdAt?: string;
}